+++ Automaton/Automaton.csproj
@@ -103,8 +103,8 @@
     <Compile Include="CompletedDialog.xaml.cs">
       <DependentUpon>CompletedDialog.xaml</DependentUpon>
     </Compile>
-    <Compile Include="GenericDialog.xaml.cs">
-      <DependentUpon>GenericDialog.xaml</DependentUpon>
+    <Compile Include="GenericErrorDialog.xaml.cs">
+      <DependentUpon>GenericErrorDialog.xaml</DependentUpon>
     </Compile>
     <Compile Include="InitialSetupDialog.xaml.cs">
       <DependentUpon>InitialSetupDialog.xaml</DependentUpon>
@@ -127,8 +127,8 @@
       <DependentUpon>OptionalsInstaller.xaml</DependentUpon>
     </Compile>
     <Compile Include="ViewModel\CompletedDialogViewModel.cs" />
-    <Compile Include="ViewModel\GenericDialogHandler.cs" />
-    <Compile Include="ViewModel\GenericDialogViewModel.cs" />
+    <Compile Include="ViewModel\GenericErrorDialogHandler.cs" />
+    <Compile Include="ViewModel\GenericErrorDialogViewModel.cs" />
     <Compile Include="ViewModel\LoadingDialogHandle.cs" />
     <Compile Include="ViewModel\LoadingDialogViewModel.cs" />
     <Compile Include="ViewModel\ScrollViewerExtensions.cs" />
@@ -142,7 +142,7 @@
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="GenericDialog.xaml">
+    <Page Include="GenericErrorDialog.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
@@ -217,6 +217,9 @@
   </ItemGroup>
   <ItemGroup>
     <None Include="Resources\7za.exe" />
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Handlers\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="..\packages\Fody.2.2.1.0\build\net452\Fody.targets" Condition="Exists('..\packages\Fody.2.2.1.0\build\net452\Fody.targets')" />
