--- Automaton/GenericDialog.xaml
@@ -1,36 +1,0 @@
-﻿<UserControl 
-            x:Class="Automaton.GenericDialog"
-            xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-            xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-            xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-            xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-            xmlns:i="clr-namespace:System.Windows.Interactivity;assembly=System.Windows.Interactivity"
-            xmlns:vm="clr-namespace:Automaton.ViewModel"
-            xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
-            TextElement.Foreground="{DynamicResource MaterialDesignBody}"
-            TextElement.FontWeight="Regular"
-            TextElement.FontSize="13"
-            TextOptions.TextFormattingMode="Ideal" 
-            TextOptions.TextRenderingMode="Auto"        
-            Background="{DynamicResource MaterialDesignPaper}"
-            FontFamily="{DynamicResource MaterialDesignFont}"
-            RenderOptions.ClearTypeHint="Enabled"
-            mc:Ignorable="d">
-    <UserControl.DataContext>
-        <vm:GenericDialogViewModel/>
-    </UserControl.DataContext>
-
-    <materialDesign:DialogHost IsOpen="{Binding IsOpen}">
-        <materialDesign:DialogHost.DialogContent>
-            <Grid Margin="5" Height="250" Width="400">
-                <StackPanel>
-                    <TextBlock FontSize="25" HorizontalAlignment="Center" />
-                    <TextBlock Foreground="Gray" HorizontalAlignment="Center"/>
-                </StackPanel>
-
-                <Button HorizontalAlignment="Right" VerticalAlignment="Bottom" Style="{DynamicResource MaterialDesignFlatButton}"
-                    Command="{Binding CloseDialogCommand}">CLOSE</Button>
-            </Grid>
-        </materialDesign:DialogHost.DialogContent>
-    </materialDesign:DialogHost>
-</UserControl>
