+++ Automaton/GenericErrorDialog.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows.Controls;
+
+namespace Automaton
+{
+    /// <summary>
+    /// Interaction logic for GenericDialog.xaml
+    /// </summary>
+    public partial class GenericErrorDialog : UserControl
+    {
+        public GenericErrorDialog()
+        {
+            InitializeComponent();
+        }
+    }
+}
