+++ Automaton/ViewModel/GenericErrorDialogHandler.cs
@@ -1,0 +1,24 @@
+﻿using GalaSoft.MvvmLight.Messaging;
+
+namespace Automaton.ViewModel
+{
+    class GenericErrorDialogHandler
+    {
+        public static void OpenDialog(string title, string errorMessage)
+        {
+            var payload = new GenericErrorDialogPayload()
+            {
+                Title = title,
+                ErrorMessage = errorMessage
+            };
+
+            Messenger.Default.Send(payload);
+        }
+    }
+
+    class GenericErrorDialogPayload
+    {
+        public string Title;
+        public string ErrorMessage;
+    }
+}
